Revamp Example Parity architecture.
- Use multi-level inheritance with interfaces.
- Introduced a common base command with shared
  functions and cmd line options.
- UI, UX fixes on frontend, etc.
